+++ src/main/java/com/todayhouse/domain/user/dao/FollowRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FollowRepository extends JpaRepository<Follow, Long> {
 
-    int countByFromId(Long id);
+    long countByFromId(Long id);
 
-    int countByToId(Long id);
+    long countByToId(Long id);
 
     void deleteByFromIdAndToId(Long fromId, Long toId);
 }
